requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-filter==22.1
django-tables2==2.4.1
sqlparse==0.4.2
tzdata==2022.1
